For deleted paths, always check for directory-level callbacks anyway
This is because isFile and isDirectory on the File objects of those paths
will always be false since the file no longer exists.

This implies that if a folder inside a folder is deleted and the original registration
was a recursive callback monitor message, then two identical callbacks will be fired.